requirements.txt
yfinance
pandas
pytz
numpy
newsapi-python
sec_api
bs4
